remove unnecessary cast to int

+++ src/main/java/g0101_0200/s0165_compare_version_numbers/Solution.java
@@ -11,7 +11,7 @@ public class Solution {
             if (c == '.') {
                 break;
             } else {
-                numA = numA * 10 + ((int) c - 48);
+                numA = numA * 10 + (c - 48);
             }
         }
 
@@ -23,7 +23,7 @@ public class Solution {
             if (c == '.') {
                 break;
             } else {
-                numB = numB * 10 + ((int) c - 48);
+                numB = numB * 10 + (c - 48);
             }
         }
